+++ lib/mutant/mutation/config.rb
@@ -3,11 +3,22 @@
 module Mutant
   class Mutation
     class Config
-      include Anima.new(:ignore_patterns, :timeout)
+      include Anima.new(
+        :ignore_patterns,
+        :operators,
+        :timeout
+      )
+
+      EMPTY = new(
+        ignore_patterns: [],
+        operators:       nil,
+        timeout:         nil
+      )
 
       DEFAULT = new(
-        timeout:         nil,
-        ignore_patterns: []
+        ignore_patterns: [],
+        operators:       Mutation::Operators::Light.new,
+        timeout:         nil
       )
 
       ignore_pattern = Transform::Block.capture('ignore pattern', &AST::Pattern.method(:parse))
@@ -21,6 +32,10 @@ module Mutant
                 value:     'ignore_patterns'
               ),
               Transform::Hash::Key.new(
+                transform: Operators::TRANSFORM,
+                value:     'operators'
+              ),
+              Transform::Hash::Key.new(
                 transform: Transform::FLOAT,
                 value:     'timeout'
               )
@@ -28,13 +43,14 @@ module Mutant
             required: []
           ),
           Transform::Hash::Symbolize.new,
-          Transform::Success.new(block: DEFAULT.method(:with))
+          Transform::Success.new(block: EMPTY.method(:with))
         ]
       )
 
       def merge(other)
         with(
           ignore_patterns: other.ignore_patterns,
+          operators:       other.operators || operators,
           timeout:         other.timeout || timeout
         )
       end