+++ spec/unit/mutant/reporter/cli/printer/env_progress_spec.rb
@@ -16,6 +16,7 @@ RSpec.describe Mutant::Reporter::CLI::Printer::EnvProgress do
         Jobs:            auto
         Includes:        []
         Requires:        []
+        Operators:       light
         Subjects:        1
         All-Tests:       2
         Available-Tests: 1
@@ -42,6 +43,7 @@ RSpec.describe Mutant::Reporter::CLI::Printer::EnvProgress do
         Jobs:            auto
         Includes:        []
         Requires:        []
+        Operators:       light
         Subjects:        1
         All-Tests:       2
         Available-Tests: 1
@@ -70,6 +72,7 @@ RSpec.describe Mutant::Reporter::CLI::Printer::EnvProgress do
         Jobs:            auto
         Includes:        []
         Requires:        []
+        Operators:       light
         Subjects:        1
         All-Tests:       2
         Available-Tests: 1